MassSectorGenerator works again
