+++ Zilon.Core/Zilon.Core.MassSectorGenerator/Program.cs
@@ -1,8 +1,5 @@
 ﻿using System;
-using System.Diagnostics;
 using System.IO;
-using System.Linq;
-using System.Reflection;
 using System.Threading.Tasks;
 
 using Microsoft.Extensions.DependencyInjection;
@@ -10,8 +7,6 @@ using Microsoft.Extensions.DependencyInjection;
 using Zilon.CommonUtilities;
 using Zilon.Core.MapGenerators;
 using Zilon.Core.Schemes;
-using Zilon.Core.Tactics;
-using Zilon.Core.Tactics.Spatial;
 using Zilon.Core.World;
 
 namespace Zilon.Core.MassSectorGenerator
@@ -23,43 +18,6 @@ namespace Zilon.Core.MassSectorGenerator
         static Program()
         {
             _random = new Random();
-        }
-
-        private static Task CheckSectorAsync(ISectorValidator[] validators, IServiceProvider scopeContainer,
-            ISector sector)
-        {
-            return Task.Run(() =>
-            {
-                var stopWatch = new Stopwatch();
-                stopWatch.Start();
-
-                var checkTasks = validators.Select(x => x.Validate(sector, scopeContainer));
-
-                var allTasks = Task.WhenAll(checkTasks);
-
-                try
-                {
-                    allTasks.Wait();
-                }
-                catch (AggregateException exception)
-                {
-                    Log.Error("Сектор содержит ошибки:");
-
-                    foreach (var inner in exception.InnerExceptions)
-                    {
-                        if (inner is SectorValidationException)
-                        {
-                            Log.Error(inner);
-                        }
-                    }
-
-                    throw;
-                }
-
-                stopWatch.Stop();
-
-                Log.Info($"CHECK DURATION: {stopWatch.Elapsed.TotalSeconds} SEC");
-            });
         }
 
         private static int GetDiceSeed(string[] args)
@@ -105,67 +63,7 @@ namespace Zilon.Core.MassSectorGenerator
             return outputPath;
         }
 
-        private static SectorSchemeResult GetSectorScheme(string[] args, ISchemeService schemeService)
-        {
-            var locationSchemeSid = ArgumentHelper.GetProgramArgument(args, Args.LOCATION_SCHEME_SID_ARG_NAME);
-            var sectorSchemeSid = ArgumentHelper.GetProgramArgument(args, Args.SECTOR_SCHEME_SID_ARG_NAME);
-            if (string.IsNullOrWhiteSpace(locationSchemeSid) && string.IsNullOrWhiteSpace(sectorSchemeSid))
-            {
-                // Если схемы не указаны, то берём случайную схему.
-                // Это используется на билд-сервере, чтобы случайно проверить несколько схем.
-
-                var locationSchemes = schemeService.GetSchemes<ILocationScheme>()
-                    .Where(x => x.SectorLevels != null && x.SectorLevels.Any())
-                    .ToArray();
-                var locationSchemeIndex = _random.Next(0, locationSchemes.Length);
-                var locationScheme = locationSchemes[locationSchemeIndex];
-
-                var sectorSchemes = locationScheme.SectorLevels;
-                var sectorSchemeIndex = _random.Next(0, sectorSchemes.Length);
-                var sectorScheme = sectorSchemes[sectorSchemeIndex];
-
-                Log.Info($"SCHEME: {locationScheme.Sid} - {sectorScheme.Sid}(index:{sectorSchemeIndex})");
-
-                var result = new SectorSchemeResult(locationScheme, sectorScheme);
-
-                return result;
-            }
-            else
-            {
-                // Если схемы заданы, то строим карту на их основе.
-                // Это будет использовано для отладки.
-
-                var locationScheme = schemeService.GetScheme<ILocationScheme>(locationSchemeSid);
-                if (locationScheme == null)
-                {
-                    throw new SectorGeneratorException($"Не найдена схема локации {locationSchemeSid}.");
-                }
-
-                var sectorScheme = locationScheme.SectorLevels.SingleOrDefault(x => x.Sid == sectorSchemeSid);
-                if (sectorScheme == null)
-                {
-                    throw new SectorGeneratorException($"Не найдена схема сектора {sectorSchemeSid}.");
-                }
-
-                var result = new SectorSchemeResult(locationScheme, sectorScheme);
-
-                return result;
-            }
-        }
-
-        private static ISectorValidator[] GetValidatorsInAssembly()
-        {
-            var thisAssembly = Assembly.GetExecutingAssembly();
-            var validatorTypes = thisAssembly.GetTypes()
-                .Where(x => typeof(ISectorValidator).IsAssignableFrom(x))
-                .Where(x => !x.IsInterface && !x.IsAbstract);
-
-            var validators = validatorTypes.Select(x => Activator.CreateInstance(x)).Cast<ISectorValidator>();
-
-            return validators.ToArray();
-        }
-
-        private static async Task Main(string[] args)
+        public static async Task Main(string[] args)
         {
             var diceSeed = GetDiceSeed(args);
             var outputPath = GetOutputPath(args);
@@ -179,39 +77,32 @@ namespace Zilon.Core.MassSectorGenerator
 
             var schemeService = serviceProvider.GetRequiredService<ISchemeService>();
 
-            var sectorSchemeResult = GetSectorScheme(args, schemeService);
+            var sectorSchemeResult = ValidationHelper.GetSectorScheme(_random, args, schemeService);
 
-            var biome = new Biome(sectorSchemeResult.Location);
-            var sectorNode = new SectorNode(biome, sectorSchemeResult.Sector);
+            var biome = new Biome(sectorSchemeResult.LocationScheme);
+            var sectorNode = new SectorNode(biome, sectorSchemeResult.SectorScheme);
+            biome.AddNode(sectorNode);
 
+            var nextCount = 3;
+            for (var i = 0; i < nextCount; i++)
+            {
+                var nextSectorNode = new SectorNode(biome, sectorSchemeResult.SectorScheme);
+                biome.AddNode(nextSectorNode);
+                biome.AddEdge(sectorNode, nextSectorNode);
+            }
+                
             var sectorFactory = serviceProvider.GetRequiredService<ISectorGenerator>();
             var sector = await sectorFactory.GenerateAsync(sectorNode).ConfigureAwait(false);
-            sector.Scheme = sectorSchemeResult.Location;
+            sector.Scheme = sectorSchemeResult.LocationScheme;
 
             // Проверка
 
-            var sectorValidators = GetValidatorsInAssembly();
-            var checkTask = CheckSectorAsync(sectorValidators, serviceProvider, sector);
+            var sectorValidators = ValidatorCollector.GetValidatorsInAssembly();
+            var checkTask = ValidationHelper.CheckSectorAsync(sectorValidators, serviceProvider, sector);
 
-            var saveTask = SaveMapAsImageAsync(outputPath, sector);
+            var saveTask = ImageHelper.SaveMapAsImageAsync(outputPath, sector);
 
             await Task.WhenAll(checkTask, saveTask).ConfigureAwait(false);
-        }
-
-        private static void SaveMapAsImage(ISectorMap map, string outputPath)
-        {
-            using var bmp = MapDrawer.DrawMap(map);
-            bmp.Save(outputPath);
-        }
-
-        private static Task SaveMapAsImageAsync(string outputPath, ISector sector)
-        {
-            if (outputPath != null)
-            {
-                SaveMapAsImage(sector.Map, outputPath);
-            }
-
-            return Task.CompletedTask;
         }
     }
 }
